Refactor Refine toolbar menu items into separate "RefineMenu" class.
