+++ pwiz_tools/Skyline/Skyline.csproj
@@ -418,6 +418,15 @@
       <DependentUpon>DocumentChangeLogEntryDlg.cs</DependentUpon>
     </Compile>
     <Compile Include="Controls\Graphs\AreaCVGraphData.cs" />
+    <Compile Include="Menus\RefineMenu.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
+    <Compile Include="Menus\RefineMenu.Designer.cs">
+      <DependentUpon>RefineMenu.cs</DependentUpon>
+    </Compile>
+    <Compile Include="Menus\SkylineControl.cs">
+      <SubType>UserControl</SubType>
+    </Compile>
     <Compile Include="Model\AbstractDdaSearchEngine.cs" />
     <Compile Include="Model\AreaCVRefinementData.cs" />
     <Compile Include="Model\AuditLog\AuditLogObject.cs" />
@@ -2508,6 +2517,9 @@
     </EmbeddedResource>
     <EmbeddedResource Include="FileUI\ChooseIrtStandardPeptidesDlg.resx">
       <DependentUpon>ChooseIrtStandardPeptidesDlg.cs</DependentUpon>
+    </EmbeddedResource>
+    <EmbeddedResource Include="Menus\RefineMenu.resx">
+      <DependentUpon>RefineMenu.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="Model\AuditLog\AuditLogStrings.ja.resx">
       <DependentUpon>AuditLogStrings.resx</DependentUpon>
